+++ src/components/candidates/Candidate.vue
