+++ src/components/questions/CategoriesList.vue
