+++ src/components/questions/__tests__/CategoriesList.spec.ts
