+++ src/views/CandidatesListView.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script setup>
-import CandidatesList from "../components/Candidates/CandidatesList.vue";
+import CandidatesList from "../components/candidates/CandidatesList.vue";
 </script>
 
 <style></style>
